Fix code-cells-forward-cell for repeated invocations
When point is just before the page delimiter,
Emacs 29+ `forward-page` just skips to the end of it.

But `code-cells-forward-cell` then "resets" the pointer
to the beginning of line which makes it impossible
to call `code-cells-forward-cell` multiple times in a row.

Moving the cursor to the next char makes it inside
the page delimiter and `forward-page` works as expected.